ATR-588: moved pxgraph create instance walker to a public class in utils
